Refactor render target, add circle mask to shader & clean stuff up a bit
