+++ code/ui/elements/weapons/ScopeRenderTarget.cs
@@ -1,0 +1,67 @@
+﻿using Sandbox;
+using Sandbox.UI;
+
+namespace Fortwars
+{
+	public class ScopeRenderTarget : Panel
+	{
+		Texture colorTexture;
+		Texture depthTexture;
+
+		public ScopeRenderTarget()
+		{
+			Vector2 textureSize = Screen.Size / 2f;
+
+			colorTexture = Texture.CreateRenderTarget()
+						 .WithSize( (int)textureSize.x, (int)textureSize.y )
+						 .WithScreenFormat()
+						 .WithScreenMultiSample()
+						 .Create();
+
+			depthTexture = Texture.CreateRenderTarget()
+						 .WithSize( (int)textureSize.x, (int)textureSize.y )
+						 .WithDepthFormat()
+						 .WithScreenMultiSample()
+						 .Create();
+		}
+
+		public override void OnDeleted()
+		{
+			colorTexture.Dispose();
+			depthTexture.Dispose();
+
+			base.OnDeleted();
+		}
+
+		public override void DrawBackground( ref RenderState state )
+		{
+			var player = Local.Pawn;
+			if ( player == null )
+				return;
+
+			if ( SceneWorld.Current == null )
+				return;
+
+			if ( player.ActiveChild is not FortwarsWeapon weapon )
+				return;
+
+			var sceneObject = weapon.ViewModelEntity.SceneObject;
+
+			if ( sceneObject == null )
+				return;
+
+			Render.DrawScene( colorTexture,
+					depthTexture,
+					new Vector2( 500, 500 ),
+					SceneWorld.Current,
+					CurrentView.Position,
+					CurrentView.Rotation.Angles(),
+					weapon.WeaponAsset.AimFovMult * 25f,
+					zNear: 1,
+					zFar: 25000 );
+
+			Render.Set( "ScopeRT", colorTexture );
+			sceneObject.SetValue( "ScopeRT", colorTexture );
+		}
+	}
+}
