+++ code/weapons/ViewModel.cs
@@ -168,6 +168,10 @@ namespace Fortwars
 			offset += CalcBobbingOffset( playerVelocity, bobCycleTime );
 			offset -= ShootOffset;
 
+			Vector2 maskOffset = new Vector2( offset.y, offset.z ) * 0.1f * (10 * offset.x + 1f);
+			DebugOverlay.ScreenText( 0, maskOffset.ToString() );
+			SceneObject.SetValue( "maskOffset", new Vector2( maskOffset.x, maskOffset.y ) );
+
 			if ( Owner.GroundEntity == null )
 			{
 				offset += new Vector3( 0, 0, -2.5f );
